remove header layout components
