+++ src/components/layout-components/overviewStatsLayout.tsx
@@ -2,15 +2,17 @@ import * as React from 'react';
 import { CompanyInfo } from '../overview/companyInfo';
 import KeyStatsContainer from '../key-stats/keyStatsContainer';
 import { TopPeers } from '../peers/topPeers';
+import { Peers, Overview } from '../../util';
 
 interface OverviewStatsProps {
-  topPeers: any;
-  overview: any;
+  topPeers: Peers;
+  overview: Overview;
 }
 
-export const OverviewStatsLayout: React.FunctionComponent<
-  OverviewStatsProps
-> = ({ overview, topPeers }: OverviewStatsProps) => {
+export const OverviewStatsLayout = ({
+  overview,
+  topPeers,
+}: OverviewStatsProps) => {
   return (
     <div className="company-wrapper">
       <KeyStatsContainer />
